자유게시판 틀 완성

+++ KOSTA_3_final/src/main/java/com/kosta/KOSTA_3_final/model/board/Board.java
@@ -1,6 +1,5 @@
 package com.kosta.KOSTA_3_final.model.board;
 
-import java.sql.Date;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -19,7 +18,6 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.kosta.KOSTA_3_final.model.subscribe.Subscribe;
 import com.kosta.KOSTA_3_final.model.user.User;
 
 
@@ -41,7 +39,7 @@ import lombok.ToString;
 public class Board {
 	@Id // 필수PK
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	int board_id;
+	Long board_id;
 	String board_title;
 	
 	@ManyToOne
