+++ KOSTA_3_final/src/main/java/com/kosta/KOSTA_3_final/model/board/BoardReply.java
@@ -35,7 +35,7 @@ import lombok.ToString;
 public class BoardReply {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	int board_reply_id;
+	Long board_reply_id;
 
 	String board_reply; //댓글내용
